[WFLY-15051] Refactored subsystem archetype to be a 'real' archetype, update to WildFly 25
