--- wildfly-subsystem-archetype/Readme.txt
@@ -1,4 +1,0 @@
-${artifactId}
-========================
-
-Generates a skeleton subsystem with tests. See the admin guide in the WildFly
