--- wildfly-subsystem-archetype/Release.txt
@@ -1,24 +1,0 @@
-To release this archetype for a new WildFly version
-
-1)change the following files:
-/pom.xml: 
--set project/version to the target WildFly version
--set project/properties/version.wildfly to the target WildFly version
-
-/archetype-pom.xml:
--set project/version/parent to the target WildFly version
--set project/version to the target WildFly version
-
-/archetype.properties:
-- set archetype.version to the target WildFly version
-
-/create-archetype.sh:
-- Update -DarchetypeVersion to the target WildFly version
-
-2) Run './create-archetype.sh install'
-
-3) Deploy the archetype by:
-$ cd target/archetype
-$ mvn deploy
-
-3) Set ../jboss-as-subsystem-example/pom.xml's  project/properties/version.wildfly to the target AS version.
