--- wildfly-subsystem-archetype/archetype-excludes.txt
@@ -1,12 +1,0 @@
-*.swp
-.classpath
-.git
-.gitignore
-.project
-.settings
-.svn
-archetype-pom.xml
-archetype-excludes.txt
-WORKLOG
-eclipse-dot-files
-release-procedure.txt
