--- wildfly-subsystem-archetype/archetype-metadata.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<archetype-descriptor
-        xsi:schemaLocation="http://maven.apache.org/plugins/maven-archetype-plugin/archetype-descriptor/1.0.0 http://maven.apache.org/xsd/archetype-descriptor-1.0.0.xsd"
-        name="jboss-as-subsystem-src"
-        xmlns="http://maven.apache.org/plugins/maven-archetype-plugin/archetype-descriptor/1.0.0"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-    <requiredProperties>
-        <requiredProperty key="name">
-            <defaultValue>WildFly subsystem project</defaultValue>
-        </requiredProperty>
-        <requiredProperty key="module"/>
-    </requiredProperties>
-    <fileSets>
-        <fileSet filtered="true" packaged="true" encoding="UTF-8">
-            <directory>src/main/java</directory>
-            <includes>
-                <include>**/*.java</include>
-            </includes>
-        </fileSet>
-        <fileSet filtered="true" encoding="UTF-8" packaged="false">
-            <directory>src/main/resources</directory>
-            <includes>
-                <include>**/*.Extension</include>
-                <include>**/module.xml</include>
-                <include>schema/*.xsd</include>
-            </includes>
-        </fileSet>
-        <fileSet filtered="true" encoding="UTF-8" packaged="true">
-            <directory>src/main/resources</directory>
-            <includes>
-                <include>**/*.properties</include>
-            </includes>
-        </fileSet>
-
-        <fileSet filtered="true" packaged="true" encoding="UTF-8">
-            <directory>src/test/java</directory>
-            <includes>
-                <include>**/*.java</include>
-            </includes>
-        </fileSet>
-
-    </fileSets>
-
-</archetype-descriptor>
