--- wildfly-subsystem-archetype/archetype-pom.xml
@@ -1,74 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.jboss</groupId>
-        <artifactId>jboss-parent</artifactId>
-        <version>35</version>
-        <relativePath />
-    </parent>
-
-    <groupId>org.wildfly.archetype</groupId>
-    <artifactId>wildfly-subsystem</artifactId>
-    <version>25.0.0.Final-SNAPSHOT</version>
-    <packaging>maven-archetype</packaging>
-
-    <name>WildFly Subsystem</name>
-
-    <description>An archetype that generates a skeleton project for implementing a WildFly subsystem</description>
-
-    <url>http://wildfly.org</url>
-
-    <scm>
-        <connection>scm:git:https://github.com/wildfly/archetypes.git</connection>
-        <developerConnection>scm:git:https://github.com/wildfly/archetypes.git</developerConnection>
-        <url>https://github.com/wildfly/archetypes/</url>
-    </scm>
-
-    <licenses>
-        <license>
-            <name>LGPL 2.1</name>
-            <url>http://www.gnu.org/licenses/lgpl-2.1.html</url>
-            <distribution>repo</distribution>
-        </license>
-    </licenses>
-
-    <developers>
-        <developer>
-            <name>Kabir Khan</name>
-            <email>kabir.khan@jboss.com</email>
-            <organization>JBoss, by Red Hat</organization>
-            <organizationUrl>http://redhat.com/jboss</organizationUrl>
-            <url>http://kabir-khan.blogspot.com/</url>
-        </developer>
-        <developer>
-            <name>Tomaz Cerar</name>
-            <email>tomaz.cerar@redhat.com</email>
-            <organization>JBoss, by Red Hat</organization>
-            <organizationUrl>http://redhat.com/jboss</organizationUrl>
-        </developer>
-    </developers>
-
-    <build>
-        <extensions>
-            <extension>
-                <groupId>org.apache.maven.archetype</groupId>
-                <artifactId>archetype-packaging</artifactId>
-                <version>2.4</version>
-            </extension>
-        </extensions>
-
-        <pluginManagement>
-            <plugins>
-                <plugin>
-                    <artifactId>maven-archetype-plugin</artifactId>
-                    <version>2.4</version>
-                    <extensions>true</extensions>
-                </plugin>
-            </plugins>
-        </pluginManagement>
-    </build>
-
-</project>
