--- wildfly-subsystem-archetype/archetype-test.properties
@@ -1,6 +1,0 @@
-#Fri Jul 08 10:36:41 BST 2011
-version=0.1-SNAPSHOT
-name=WildFly subsystem project
-groupId=archetype.it
-artifactId=basic
-module=org.blah
