--- wildfly-subsystem-archetype/archetype.properties
@@ -1,9 +1,0 @@
-archetype.groupId=org.wildfly.archetype
-archetype.artifactId=wildfly-subsystem
-archetype.version=25.0.0.Final-SNAPSHOT
-# Java package of the source files; will be replaced with the ${package} expression
-archetype.package=com.mycompany.subsystem
-archetype.languages=java, properties
-# value of <artifactId> element in POM is converted to ${artifactId} expression in filtered files
-archetype.filteredExtensions=java,txt,html,properties
-name=WildFly subsystem project
