--- wildfly-subsystem-archetype/src/main/java/com/mycompany/subsystem/extension/SubsystemExtension.java
@@ -1,103 +1,0 @@
-package com.mycompany.subsystem.extension;
-
-import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.SUBSYSTEM;
-
-import java.util.List;
-
-import javax.xml.stream.XMLStreamConstants;
-import javax.xml.stream.XMLStreamException;
-
-import org.jboss.as.controller.Extension;
-import org.jboss.as.controller.ExtensionContext;
-import org.jboss.as.controller.PathAddress;
-import org.jboss.as.controller.PathElement;
-import org.jboss.as.controller.PersistentResourceXMLDescription;
-import org.jboss.as.controller.SubsystemRegistration;
-import org.jboss.as.controller.descriptions.StandardResourceDescriptionResolver;
-import org.jboss.as.controller.operations.common.GenericSubsystemDescribeHandler;
-import org.jboss.as.controller.parsing.ExtensionParsingContext;
-import org.jboss.as.controller.persistence.SubsystemMarshallingContext;
-import org.jboss.as.controller.registry.ManagementResourceRegistration;
-import org.jboss.dmr.ModelNode;
-import org.jboss.staxmapper.XMLElementReader;
-import org.jboss.staxmapper.XMLElementWriter;
-import org.jboss.staxmapper.XMLExtendedStreamReader;
-import org.jboss.staxmapper.XMLExtendedStreamWriter;
-
-
-/**
- * @author <a href="kabir.khan@jboss.com">Kabir Khan</a>
- */
-public class SubsystemExtension implements Extension {
-
-    /**
-     * The name space used for the {@code substystem} element
-     */
-    public static final String NAMESPACE = "urn:mycompany:mysubsystem:1.0";
-
-    /**
-     * The name of our subsystem within the model.
-     */
-    public static final String SUBSYSTEM_NAME = "mysubsystem";
-
-    /**
-     * The parser used for parsing our subsystem
-     */
-    private final SubsystemParser parser = new SubsystemParser();
-
-    protected static final PathElement SUBSYSTEM_PATH = PathElement.pathElement(SUBSYSTEM, SUBSYSTEM_NAME);
-    private static final String RESOURCE_NAME = SubsystemExtension.class.getPackage().getName() + ".LocalDescriptions";
-
-    static StandardResourceDescriptionResolver getResourceDescriptionResolver(final String keyPrefix) {
-        String prefix = SUBSYSTEM_NAME + (keyPrefix == null ? "" : "." + keyPrefix);
-        return new StandardResourceDescriptionResolver(prefix, RESOURCE_NAME, SubsystemExtension.class.getClassLoader(), true, false);
-    }
-
-    @Override
-    public void initializeParsers(ExtensionParsingContext context) {
-        context.setSubsystemXmlMapping(SUBSYSTEM_NAME, NAMESPACE, parser);
-    }
-
-
-    @Override
-    public void initialize(ExtensionContext context) {
-        final SubsystemRegistration subsystem = context.registerSubsystem(SUBSYSTEM_NAME, 1, 0);
-        final ManagementResourceRegistration registration = subsystem.registerSubsystemModel(SubsystemDefinition.INSTANCE);
-        registration.registerOperationHandler(GenericSubsystemDescribeHandler.DEFINITION, GenericSubsystemDescribeHandler.INSTANCE);
-
-        subsystem.registerXMLElementWriter(parser);
-    }
-
-    /**
-     * The subsystem parser, which uses stax to read and write to and from xml
-     */
-    private static class SubsystemParser implements XMLStreamConstants, XMLElementReader<List<ModelNode>>, XMLElementWriter<SubsystemMarshallingContext> {
-
-        private final PersistentResourceXMLDescription xmlDescription;
-
-        private SubsystemParser() {
-            this.xmlDescription = PersistentResourceXMLDescription.builder(SubsystemDefinition.INSTANCE)
-                    // use .addChild(...) to add any children
-                    .build();
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        public void writeContent(XMLExtendedStreamWriter writer, SubsystemMarshallingContext context) throws XMLStreamException {
-            ModelNode model = new ModelNode();
-            model.get(SubsystemDefinition.INSTANCE.getPathElement().getKeyValuePair()).set(context.getModelNode());
-            xmlDescription.persist(writer, model, SubsystemExtension.NAMESPACE);
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        public void readElement(XMLExtendedStreamReader reader, List<ModelNode> list) throws XMLStreamException {
-            xmlDescription.parse(reader, PathAddress.EMPTY_ADDRESS, list);
-        }
-    }
-
-}
