--- wildfly-subsystem-archetype/src/main/java/com/mycompany/subsystem/extension/deployment/SubsystemDeploymentProcessor.java
@@ -1,43 +1,0 @@
-package com.mycompany.subsystem.extension.deployment;
-
-import org.jboss.as.server.AbstractDeploymentChainStep;
-import org.jboss.as.server.deployment.DeploymentPhaseContext;
-import org.jboss.as.server.deployment.DeploymentUnit;
-import org.jboss.as.server.deployment.DeploymentUnitProcessingException;
-import org.jboss.as.server.deployment.DeploymentUnitProcessor;
-import org.jboss.as.server.deployment.Phase;
-import org.jboss.logging.Logger;
-
-/**
- * An example deployment unit processor that does nothing. To add more deployment
- * processors copy this class, and add to the {@link AbstractDeploymentChainStep}
- * {@link com.mycompany.subsystem.extension.SubsystemAdd#performBoottime(org.jboss.as.controller.OperationContext, org.jboss.dmr.ModelNode, org.jboss.dmr.ModelNode, org.jboss.as.controller.ServiceVerificationHandler, java.util.List)}
- *
- * @author <a href="kabir.khan@jboss.com">Kabir Khan</a>
- */
-public class SubsystemDeploymentProcessor implements DeploymentUnitProcessor {
-
-    Logger log = Logger.getLogger(SubsystemDeploymentProcessor.class);
-
-    /**
-     * See {@link Phase} for a description of the different phases
-     */
-    public static final Phase PHASE = Phase.DEPENDENCIES;
-
-    /**
-     * The relative order of this processor within the {@link #PHASE}.
-     * The current number is large enough for it to happen after all
-     * the standard deployment unit processors that come with JBoss AS.
-     */
-    public static final int PRIORITY = 0x4000;
-
-    @Override
-    public void deploy(DeploymentPhaseContext phaseContext) throws DeploymentUnitProcessingException {
-        log.info("Deploy");
-    }
-
-    @Override
-    public void undeploy(DeploymentUnit context) {
-    }
-
-}
