+++ wildfly-subsystem-archetype/src/main/resources/META-INF/maven/archetype-metadata.xml
@@ -1,0 +1,66 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Licensed under the Apache License, Version 2.0 (the "License"); you 
+    may not use this file except in compliance with the License. You may obtain 
+    a copy of the License at http://www.apache.org/licenses/LICENSE-2.0 Unless 
+    required by applicable law or agreed to in writing, software distributed 
+    under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES 
+    OR CONDITIONS OF ANY KIND, either express or implied. See the License for 
+    the specific language governing permissions and limitations under the License. -->
+<archetype-descriptor name="wildfly-jakartaee-webapp-archetype">
+
+    <requiredProperties>
+        <requiredProperty key="name">
+            <defaultValue>WildFly subsystem project</defaultValue>
+        </requiredProperty>
+        <requiredProperty key="module"/>
+    </requiredProperties>
+	
+    <fileSets>
+        <fileSet filtered="true" packaged="false">
+            <directory></directory>
+            <includes>
+                <include>README.txt</include>
+            </includes>
+        </fileSet>
+        <fileSet filtered="false" packaged="false">
+            <directory></directory>
+            <includes>
+                <include>.gitignore</include>
+            </includes>
+        </fileSet>
+        
+        <fileSet filtered="true" packaged="true" encoding="UTF-8">
+            <directory>src/main/java</directory>
+        </fileSet>
+        <fileSet encoding="UTF-8" packaged="true" filtered="true">
+            <directory>src/test/java</directory>
+        </fileSet>
+        <fileSet encoding="UTF-8">
+            <directory>src/main/webapp</directory>
+        </fileSet>
+        <fileSet encoding="UTF-8" filtered="true" packaged="true">
+            <!--"LocalDescriptions.properties" is placed in a package directory structure-->
+            <directory>src/main/resources</directory>
+            <includes>
+                <include>LocalDescriptions.properties</include>
+            </includes>
+        </fileSet>
+		<!--Non-packaged filed: -->
+		<fileSet encoding="UTF-8" filtered="true" packaged="false">
+            <!--"LocalDescriptions.properties" is placed in a package directory structure-->
+            <directory>src/main/resources</directory>
+            <includes>
+                <include>META-INF/**/*.*</include>
+				<include>module/**/*.*</include>
+				<include>schema/*.*</include>
+            </includes>
+        </fileSet>
+        <fileSet encoding="UTF-8" filtered="true">
+            <directory>src/test/resources</directory>
+            <includes>
+                <include>**/*.xml</include>
+            </includes>
+        </fileSet>
+    </fileSets>
+
+</archetype-descriptor>
