--- wildfly-subsystem-archetype/src/main/resources/META-INF/services/org.jboss.as.controller.Extension
@@ -1,4 +1,0 @@
-#set( $symbol_pound = '#' )
-#set( $symbol_dollar = '$' )
-#set( $symbol_escape = '\' )
-${package}.extension.SubsystemExtension
