+++ wildfly-subsystem-archetype/src/main/resources/archetype-resources/src/main/java/SubsystemRemove.java
@@ -1,0 +1,34 @@
+package ${package};
+
+import org.jboss.as.controller.AbstractRemoveStepHandler;
+import org.jboss.as.controller.OperationContext;
+import org.jboss.as.controller.OperationFailedException;
+import org.jboss.dmr.ModelNode;
+
+/**
+ * Handler responsible for removing the subsystem resource from the model
+ *
+ * @author <a href="kabir.khan@jboss.com">Kabir Khan</a>
+ */
+class SubsystemRemove extends AbstractRemoveStepHandler {
+
+    static final SubsystemRemove INSTANCE = new SubsystemRemove();
+
+
+    private SubsystemRemove() {
+    }
+
+    @Override
+    protected void performRuntime(OperationContext context, ModelNode operation, ModelNode model) throws OperationFailedException {
+        //Remove any services installed by the corresponding add handler here
+        //context.removeService(ServiceName.of("some", "name"));
+    }
+
+    @Override
+    protected void recoverServices(OperationContext context, ModelNode operation, ModelNode model) throws OperationFailedException {
+        // Restore any service removed in performRuntime by calling the same method the add handler uses.
+        SubsystemAdd.installServices(context, operation, model);
+    }
+
+
+}
