+++ wildfly-subsystem-archetype/src/main/resources/archetype-resources/src/main/java/deployment/SubsystemDeploymentProcessor.java
@@ -1,0 +1,44 @@
+#set( $hash = '#' )
+package ${package}.deployment;
+
+import org.jboss.as.server.AbstractDeploymentChainStep;
+import org.jboss.as.server.deployment.DeploymentPhaseContext;
+import org.jboss.as.server.deployment.DeploymentUnit;
+import org.jboss.as.server.deployment.DeploymentUnitProcessingException;
+import org.jboss.as.server.deployment.DeploymentUnitProcessor;
+import org.jboss.as.server.deployment.Phase;
+import org.jboss.logging.Logger;
+
+/**
+ * An example deployment unit processor that does nothing. To add more deployment
+ * processors copy this class, and add to the {@link AbstractDeploymentChainStep}
+ * {@link ${package}.SubsystemAdd${hash}performBoottime(org.jboss.as.controller.OperationContext, org.jboss.dmr.ModelNode, org.jboss.as.controller.registry.Resource)}
+ *
+ * @author <a href="kabir.khan@jboss.com">Kabir Khan</a>
+ */
+public class SubsystemDeploymentProcessor implements DeploymentUnitProcessor {
+
+    Logger log = Logger.getLogger(SubsystemDeploymentProcessor.class);
+
+    /**
+     * See {@link Phase} for a description of the different phases
+     */
+    public static final Phase PHASE = Phase.DEPENDENCIES;
+
+    /**
+     * The relative order of this processor within the {@link #PHASE}.
+     * The current number is large enough for it to happen after all
+     * the standard deployment unit processors that come with JBoss AS.
+     */
+    public static final int PRIORITY = 0x4000;
+
+    @Override
+    public void deploy(DeploymentPhaseContext phaseContext) throws DeploymentUnitProcessingException {
+        log.info("Deploy");
+    }
+
+    @Override
+    public void undeploy(DeploymentUnit context) {
+    }
+
+}
