+++ wildfly-subsystem-archetype/src/main/resources/archetype-resources/src/main/resources/LocalDescriptions.properties
