+++ wildfly-subsystem-archetype/src/main/resources/archetype-resources/src/main/resources/META-INF/services/org.jboss.as.controller.Extension
@@ -1,0 +1,1 @@
+${package}.SubsystemExtension
