--- wildfly-subsystem-archetype/src/main/resources/module/main/module.xml
@@ -1,19 +1,0 @@
-#set( $symbol_pound = '#' )
-#set( $symbol_dollar = '$' )
-#set( $symbol_escape = '\' )
-<module xmlns="urn:jboss:module:1.0" name="${module}">
-    <resources>
-        <resource-root path="${artifactId}.jar"/>
-    </resources>
-    
-    <dependencies>
-        <module name="javax.api"/>
-        <module name="org.jboss.staxmapper"/>
-        <module name="org.jboss.as.controller"/>
-        <module name="org.jboss.as.server"/>
-        <module name="org.jboss.modules"/>
-        <module name="org.jboss.msc"/>
-        <module name="org.jboss.logging"/>
-        <module name="org.jboss.vfs"/>
-    </dependencies>
-</module>
