+++ wildfly-subsystem-archetype/src/test/resources/projects/multi/archetype.properties
@@ -1,0 +1,6 @@
+version=0.1-SNAPSHOT
+name=WildFly subsystem project
+groupId=com.acme
+artifactId=example-subsystem
+module=org.test.subsystem
+package=com.acme.example
