+++ wildfly-subsystem-archetype/src/test/resources/projects/multi/goal.txt
@@ -1,0 +1,1 @@
+verify
